refactor : ResultView 생성 및 패키지 구조 변경
before
1. Racing 및 Main 에서 Result 출력 중
2. 예외, 랜덤 숫자 생성기 등의 패키지 위치 모호함

after
1. 결과 출력 UI 분리를 위해 ResultView 생성 - 아직 출력 위치는 변경되지 않음
2. 예외 패키지, 유틸 패키지로 리팩토링 진행
