+++ src/main/java/racingcar/domain/Car.java
@@ -1,5 +1,7 @@
 package racingcar.domain;
 
+import racingcar.util.RandomGenerator;
+
 public class Car {
     private final String name;
     private int position;
