+++ src/main/java/racingcar/view/InputView.java
@@ -1,5 +1,7 @@
 package racingcar.view;
 
+import racingcar.exception.InvalidInputException;
+
 import java.util.Arrays;
 import java.util.HashSet;
 import java.util.Scanner;
