+++ src/main/java/racingcar/view/ResultView.java
@@ -1,0 +1,4 @@
+package racingcar.view;
+
+public class ResultView {
+}
